Move PtrAs and Integer to relay package
